requirements.txt
Django==4.1.3
djangorestframework
django-rest-passwordreset
requests
pyyaml
ujson
django-filter
pytest
pytest-django
model_bakery
drf-spectacular==0.21.2
drf-spectacular-sidecar==2022.3.14
django-allauth